fix: small changes to get the export to work properly
