+++ flashcard-roguelike/game/entity/dungeon_generator/rooms/treasure_room.tscn
@@ -6,11 +6,11 @@
 [ext_resource type="Texture2D" uid="uid://dexvas5w2y7xh" path="res://assets/models/treasure_room/shelf_room/Shelves_Base_Color.png" id="5_kwurh"]
 [ext_resource type="PackedScene" uid="uid://c175n268mh5vw" path="res://game/entity/light/floating_light.tscn" id="6_04r8m"]
 [ext_resource type="PackedScene" uid="uid://cm0r6jfidsyc8" path="res://assets/models/desk/desk.glb" id="6_e5evu"]
-[ext_resource type="Material" uid="uid://cicbhcj8wuo0d" path="res://assets/models/desk/desk_texture.tres" id="7_0ufai"]
-[ext_resource type="Material" uid="uid://c6pv2i4awlv7m" path="res://assets/models/big_pencil/pencil_texture.tres" id="8_jlue3"]
+[ext_resource type="Material" path="res://assets/models/desk/desk_texture.tres" id="7_0ufai"]
+[ext_resource type="Material" path="res://assets/models/big_pencil/pencil_texture.tres" id="8_jlue3"]
 [ext_resource type="Shader" uid="uid://dq2h2yqhsu8fl" path="res://game/entity/dungeon_generator/rooms/event_rooms/shooting_event_room.gdshader" id="9_ltfas"]
 [ext_resource type="Texture2D" uid="uid://b0gi538rd7txo" path="res://assets/backgrounds/prettysleepy-stars-3750824.png" id="10_y1f5v"]
-[ext_resource type="PackedScene" uid="uid://dxdxbn1isdu44" path="res://game/entity/treasure/treasure_chest_new.tscn" id="12_0ufai"]
+[ext_resource type="PackedScene" path="res://game/entity/treasure/treasure_chest_new.tscn" id="12_0ufai"]
 
 [sub_resource type="StandardMaterial3D" id="StandardMaterial3D_wtroa"]
 resource_name = "Material.009"
