+++ flashcard-roguelike/game/entity/item/item_resources/stats/blood_vial.tres
@@ -1,10 +1,10 @@
 [gd_resource type="Resource" script_class="ItemResource" format=3 uid="uid://dhlcwfdiu3op"]
 
-[ext_resource type="Texture2D" uid="uid://ddjesu2w8nqqx" path="res://assets/sprites/item_sprites/blood_vial.png" id="1_4r2bu"]
-[ext_resource type="Script" uid="uid://3d8wv0hhafrq" path="res://game/entity/item/effects/mod_components/ModHealth.cs" id="2_4r2bu"]
+[ext_resource type="Texture2D" uid="uid://dkxdnd3h2p1nl" path="res://assets/sprites/item_sprites/blood_vial.png" id="1_4r2bu"]
+[ext_resource type="Script" uid="uid://cslf8mk74f83c" path="res://game/entity/item/effects/mod_components/ModHealth.cs" id="2_4r2bu"]
 [ext_resource type="PackedScene" uid="uid://bxt23w0064a0j" path="res://assets/models/item_models/blood_vial.glb" id="3_4r2bu"]
-[ext_resource type="Script" uid="uid://otpise5by8no" path="res://game/entity/item/effects/mod_components/ModAttack.cs" id="3_dn0k7"]
-[ext_resource type="Script" uid="uid://bmb23qe3hj2fc" path="res://shared/resources/ItemResource.cs" id="4_wem3c"]
+[ext_resource type="Script" uid="uid://crv315mdj71bb" path="res://game/entity/item/effects/mod_components/ModAttack.cs" id="3_dn0k7"]
+[ext_resource type="Script" uid="uid://573w4ti026as" path="res://shared/resources/ItemResource.cs" id="4_wem3c"]
 
 [sub_resource type="Resource" id="Resource_4obue"]
 script = ExtResource("2_4r2bu")
