+++ flashcard-roguelike/game/entity/item/item_resources/stats/iron_core.tres
@@ -1,10 +1,10 @@
 [gd_resource type="Resource" script_class="ItemResource" format=3 uid="uid://btfjb1kjvo6oi"]
 
-[ext_resource type="Texture2D" uid="uid://dpag66os7sxf" path="res://assets/sprites/item_sprites/iron_core.png" id="1_frwcj"]
-[ext_resource type="Script" uid="uid://3d8wv0hhafrq" path="res://game/entity/item/effects/mod_components/ModHealth.cs" id="2_a4r3n"]
-[ext_resource type="PackedScene" uid="uid://s21gnxl5jpo5" path="res://assets/models/item_models/iron_core.glb" id="2_a4wu3"]
-[ext_resource type="Script" uid="uid://otpise5by8no" path="res://game/entity/item/effects/mod_components/ModAttack.cs" id="2_frwcj"]
-[ext_resource type="Script" uid="uid://bmb23qe3hj2fc" path="res://shared/resources/ItemResource.cs" id="4_tw5yo"]
+[ext_resource type="Texture2D" uid="uid://sf3splvr0oq0" path="res://assets/sprites/item_sprites/iron_core.png" id="1_frwcj"]
+[ext_resource type="Script" uid="uid://cslf8mk74f83c" path="res://game/entity/item/effects/mod_components/ModHealth.cs" id="2_a4r3n"]
+[ext_resource type="PackedScene" uid="uid://hyiq0y77tq3y" path="res://assets/models/item_models/iron_core.glb" id="2_a4wu3"]
+[ext_resource type="Script" uid="uid://crv315mdj71bb" path="res://game/entity/item/effects/mod_components/ModAttack.cs" id="2_frwcj"]
+[ext_resource type="Script" uid="uid://573w4ti026as" path="res://shared/resources/ItemResource.cs" id="4_tw5yo"]
 
 [sub_resource type="Resource" id="Resource_frwcj"]
 script = ExtResource("2_a4r3n")
