+++ flashcard-roguelike/game/entity/item/item_resources/tool/leeching_charm.tres
@@ -1,11 +1,11 @@
 [gd_resource type="Resource" script_class="ItemResource" format=3 uid="uid://ce6e1who3gvtw"]
 
-[ext_resource type="Texture2D" uid="uid://jwksg1ut0kjd" path="res://assets/sprites/item_sprites/leet.png" id="1_c61nb"]
+[ext_resource type="Texture2D" uid="uid://ca3m2axlvykvc" path="res://assets/sprites/item_sprites/leet.png" id="1_c61nb"]
 [ext_resource type="PackedScene" uid="uid://ca3c6odggg1os" path="res://assets/models/item_models/leek.glb" id="2_wumle"]
-[ext_resource type="Script" uid="uid://bfc4332sy4op3" path="res://game/entity/item/ItemEffect.cs" id="3_f23tc"]
-[ext_resource type="Script" uid="uid://bmb23qe3hj2fc" path="res://shared/resources/ItemResource.cs" id="3_wumle"]
-[ext_resource type="Script" uid="uid://3d8wv0hhafrq" path="res://game/entity/item/effects/mod_components/ModHealth.cs" id="4_blujg"]
-[ext_resource type="Script" uid="uid://dgmlsi5dqsy70" path="res://game/entity/item/effects/mod_battle/OnAttackEffect.cs" id="5_d1kea"]
+[ext_resource type="Script" uid="uid://dnm5daublmlwi" path="res://game/entity/item/ItemEffect.cs" id="3_f23tc"]
+[ext_resource type="Script" uid="uid://573w4ti026as" path="res://shared/resources/ItemResource.cs" id="3_wumle"]
+[ext_resource type="Script" uid="uid://cslf8mk74f83c" path="res://game/entity/item/effects/mod_components/ModHealth.cs" id="4_blujg"]
+[ext_resource type="Script" uid="uid://c83nrwbwfv6bh" path="res://game/entity/item/effects/mod_battle/OnAttackEffect.cs" id="5_d1kea"]
 
 [sub_resource type="Resource" id="Resource_f23tc"]
 script = ExtResource("4_blujg")
