+++ flashcard-roguelike/game/entity/item/item_resources/use/health.tres
@@ -1,7 +1,7 @@
 [gd_resource type="Resource" script_class="ItemResource" format=3 uid="uid://d1k3ywv6qs5j8"]
 
 [ext_resource type="Texture2D" uid="uid://d4d7wgexibub5" path="res://assets/sprites/item_sprites/heart.png" id="1_mwtb2"]
-[ext_resource type="PackedScene" uid="uid://biwgii6xmmi1d" path="res://assets/models/item_models/heart.glb" id="1_uxjkd"]
+[ext_resource type="PackedScene" uid="uid://biwgii6xmmi1d" path="res://assets/models/item_models/Heart.glb" id="1_uxjkd"]
 [ext_resource type="Script" uid="uid://cslf8mk74f83c" path="res://game/entity/item/effects/mod_components/ModHealth.cs" id="3_3bnc8"]
 [ext_resource type="Script" uid="uid://573w4ti026as" path="res://shared/resources/ItemResource.cs" id="4_fm70t"]
