+++ flashcard-roguelike/shared/resources/ItemDatabase.cs
@@ -1,0 +1,9 @@
+using Godot;
+using Godot.Collections;
+
+[GlobalClass]
+public partial class ItemDatabase : Resource
+{
+    [Export]
+    public Array<ItemResource> Items = new();
+}
